OTA clean up
