+++ src/screens/Settings/about.tsx
@@ -9,7 +9,7 @@ import usePatrons from '../../api/queries/patrons'
 import { useInfoCaption } from '../../hooks/use-caption'
 import { version } from '../../../package.json'
 import { getStoredOtaVersion } from 'react-native-nitro-ota'
-import { downloadUpdate } from '../../components/OtaUpdates'
+import { downloadUpdate } from '../../services/ota'
 import { ICON_PRESS_STYLES } from '../../configs/style.config'
 
 function PatronsList({ patrons }: { patrons: { fullName: string }[] | undefined }) {
